Kamigaku branch

+++ core/src/com/kamigaku/dungeoncrawler/tile/Tile.java
@@ -6,14 +6,13 @@ import com.kamigaku.dungeoncrawler.component.PhysicsComponent;
 import com.kamigaku.dungeoncrawler.listener.CollisionListener;
 
 
-// IN WORK => a mettre dans AMapEntity
 public abstract class Tile implements CollisionListener {
    
     protected PhysicsComponent _physics;
     protected GraphicsComponent _graphics;
     public boolean isCrossable;
-    public float x;
-    public float y;
+    public int x;
+    public int y;
     
     public PhysicsComponent getPhysicsComponent() {
         return this._physics;
