+++ core/src/com/kamigaku/dungeoncrawler/tile/Wall.java
@@ -11,7 +11,7 @@ import com.kamigaku.dungeoncrawler.constants.Constants;
 
 public class Wall extends Tile {
 
-    public Wall(Sprite sprite, float x, float y) {
+    public Wall(Sprite sprite, int x, int y) {
         this._graphics = new GraphicsComponent(sprite, -0f, 0);
         this._physics = new PhysicsComponent(x, y, BodyType.StaticBody, Constants.CATEGORY_SCENERY,
                         (short) (Constants.CATEGORY_PLAYER | Constants.CATEGORY_MONSTER), 16f, 16f);
@@ -21,7 +21,7 @@ public class Wall extends Tile {
         this.isCrossable = false;
     }
     
-    public Wall(String sprite, float x, float y) {
+    public Wall(String sprite, int x, int y) {
         this._graphics = new GraphicsComponent(sprite, -0f, 0f);
         this._physics = new PhysicsComponent(x, y, BodyType.StaticBody, Constants.CATEGORY_SCENERY,
                         (short) (Constants.CATEGORY_PLAYER | Constants.CATEGORY_MONSTER), 16f, 16f);
